+++ apps/designmanual-frontend/app/root/layout/RootLayout.tsx
@@ -1,12 +1,12 @@
-import { Box, Flex } from "@chakra-ui/react";
+import { Flex } from "@chakra-ui/react";
 import { useEffect, useState } from "react";
 import { useLocation } from "react-router";
 
-import { useStickymenu } from "~/routes/_base/content-menu/utils";
 import { LeftSidebar } from "~/routes/_base/left-sidebar/LeftSidebar";
 import { sendPageViewEvent } from "~/utils/analytics/metabaseCore";
 
 import { Footer } from "./Footer";
+import { HeaderOffsetContext } from "./HeaderOffsetContext";
 import { SiteHeader } from "./SiteHeader";
 
 type BaseLayoutProps = {
@@ -25,50 +25,37 @@ function usePageTracking() {
   }, [location.pathname]);
 }
 export const RootLayout = ({ children }: BaseLayoutProps) => {
-  const [headerOffset, setHeaderOffset] = useState(0);
-  const { asideRef, forceFixed, fixedRect } = useStickymenu();
+  const [headerOffset, setHeaderOffset] = useState(110);
 
   usePageTracking();
 
   return (
-    <Flex direction="column" minHeight="100vh" bg="bg" fontFamily="Vy Sans">
-      <SiteHeader onHeightChange={setHeaderOffset} />
+    <HeaderOffsetContext value={headerOffset}>
+      <Flex direction="column" minHeight="100vh" bg="bg" fontFamily="Vy Sans">
+        <SiteHeader onHeightChange={setHeaderOffset} />
 
-      <Flex
-        id="content"
-        justifyContent="space-between"
-        gap={8}
-        marginX={{ base: "4", md: "8" }}
-        overflow="visible"
-        flexDirection={["column", null, null, "row"]}
-        flex={1}
-      >
-        {forceFixed && fixedRect && (
-          <Box
-            width={`${fixedRect.width}px`}
-            height={`${fixedRect.height}px`}
-            as="div"
-          />
-        )}
-        <Box
-          ref={asideRef}
-          alignSelf="flex-start"
-          position={forceFixed ? "fixed" : "sticky"}
-          top={headerOffset}
-          as="aside"
-          style={
-            forceFixed && fixedRect
-              ? { left: `${fixedRect.left}px`, width: `${fixedRect.width}px` }
-              : undefined
-          }
-          transition="all .3s linear"
+        <Flex
+          marginX={[2, 6, 8, 0]}
+          marginRight={[2, 6, 6, 6]}
+          flex={1}
+          position="relative"
+          minWidth={0}
         >
-          <LeftSidebar />
-        </Box>
-
-        {children}
+          <LeftSidebar headerOffset={headerOffset} />
+          {/* Add left margin on large screens to account for the fixed sidebar width (20rem) */}
+          <Flex
+            as="main"
+            alignItems="stretch"
+            marginLeft={[0, null, null, "21rem"]}
+            paddingTop={8}
+            flex={1}
+            minWidth={0}
+          >
+            {children}
+          </Flex>
+        </Flex>
+        <Footer />
       </Flex>
-      <Footer />
-    </Flex>
+    </HeaderOffsetContext>
   );
 };