+++ apps/designmanual-frontend/app/routes/_base.$section.$page/route.tsx
@@ -1,5 +1,5 @@
 import { groq } from "@sanity/groq-store";
-import { Flex } from "@vygruppen/spor-react";
+import { Box, Flex } from "@vygruppen/spor-react";
 import { LoaderFunctionArgs, useLoaderData } from "react-router";
 import invariant from "tiny-invariant";
 
@@ -19,7 +19,10 @@ import {
   resolveVideoPlayerGroq,
 } from "~/features/cms/sanity/query";
 import { PortableText } from "~/features/portable-text/PortableText";
+import { useHeaderOffset } from "~/root/layout/HeaderOffsetContext";
 import { getClient } from "~/utils/sanity/client";
+
+import { RightSidebar } from "../_base/right-sidebar/RightSidebar";
 
 export const loader = async ({ request, params }: LoaderFunctionArgs) => {
   invariant(params.section, "Expected params.section");
@@ -71,14 +74,47 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
 
 export default function Index() {
   const { data } = useLoaderData<typeof loader>();
+  const headerOffset = useHeaderOffset();
 
   if (!data) {
     return null;
   }
 
   return (
-    <Flex as="main" flex="1" flexDirection="column" gap={2}>
-      <PortableText value={data?.content} />
+    <Flex gap={5} justifyContent="space-between">
+      <Box flex={1} minWidth={0}>
+        <Flex as="main" flexDirection="column" gap={2}>
+          <PortableText value={data?.content} />
+        </Flex>
+        <Box
+          width="20%"
+          display="none"
+          position="fixed"
+          overflow="auto"
+          right={0}
+          paddingLeft={1}
+          paddingTop={3}
+          top={`${headerOffset}px`}
+          transition="all .3s linear"
+          height={`calc(100vh - ${headerOffset}px)`}
+          css={{
+            [`@media screen and (min-width: 1110px)`]: {
+              display: "block",
+            },
+          }}
+        >
+          <RightSidebar />
+        </Box>
+      </Box>
+      <Box
+        width="20%"
+        display="none"
+        css={{
+          [`@media screen and (min-width: 1110px)`]: {
+            display: "block",
+          },
+        }}
+      />
     </Flex>
   );
 }