+++ packages/spor-mcp-server/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @vygruppen/spor-mcp-server
 
+## 2.1.10
+
+### Patch Changes
+
+- Updated dependencies [6412d40]
+- Updated dependencies [e864f09]
+- Updated dependencies [17a00b7]
+  - @vygruppen/spor-react@13.4.0
+
+## 2.1.9
+
+### Patch Changes
+
+- Updated dependencies [b1e5186]
+- Updated dependencies [d6eddea]
+- Updated dependencies [0a35c66]
+  - @vygruppen/spor-react@13.3.2
+  - @vygruppen/spor-design-tokens@5.0.4
+
+## 2.1.8
+
+### Patch Changes
+
+- Updated dependencies [55e243a]
+- Updated dependencies [b365786]
+  - @vygruppen/spor-react@13.3.1
+
 ## 2.1.7
 
 ### Patch Changes